requirements.txt
tqdm >= 4.0.0
google-cloud-bigquery==2.3.1
Flask==1.1.2
python-dateutil >= 2.8.0
pandas >= 1.1.0
pyarrow >= 2.0.0

